Move playground code out of Select class' folder
